requirements.txt
Flask==2.2.5
Flask-Admin==1.6.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
Flask-Migrate==4.0.4
Pillow==9.5.0
Flask-CKEditor==0.4.6
Flask-Mail==0.9.1
Werkzeug==2.2.2
Flask-Cors==4.0.0
gunicorn
